Several moving window functions

These functions do not handle NA values

mov_mean(
  data,
  window_size,
  type = c("ogita", "normal", "weighted", "fading"),
  eps = 0.9
)

mov_var(
  data,
  window_size,
  type = c("ogita", "normal", "weighted", "fading"),
  eps = 0.9
)

mov_sum(
  data,
  window_size,
  type = c("ogita", "normal", "weighted", "fading"),
  eps = 0.9
)

mov_max(data, window_size)

mov_min(data, window_size)

mov_std(data, window_size, rcpp = TRUE)

movmean_std(data, window_size, rcpp = TRUE)

muinvn(data, window_size, rcpp = TRUE, n_workers = 1)

Arguments

data
A vector or a column matrix of numeric.

window_size
An integer. The size of the rolling window.

type
A string. Select between several algorithms. Default is ogita (See details).

eps
A numeric. Used only for fading algorithms (See details), otherwise has no effect.

rcpp
A logical. If TRUE will use the Rcpp implementation, otherwise will use the R implementation, that may or not be slower.

n_workers
An integer. The number of threads using for computing. Defaults to 1.

Value

mov_mean() returns a vector with moving avg.

mov_var() returns a vector with moving var.

mov_sum() returns a vector with moving sum.

mov_max() returns a vector with moving max.

mov_min() returns a vector with moving min.

mov_std() returns a vector with moving sd.

movmean_std() returns a list with vectors of the moving avg, sd, sig, sum and sqrsum.

muinvn() returns a list with vectors of moving avg and sig.

Details

Some functions may use different algorithms to compute the results. The available types are:

ogita: This is the default. It uses the Ogita et al., Accurate Sum, and Dot Product for precision. It is not the fastest algorithm, but the time spent vs. guarantee of precision worth it.

normal: This uses the cumsum method that is faster, but unreliable in some situations (I have to find the references, but is true).

weighted: This uses Rodrigues P., et al. algorithm that uses a weighted window for online purposes. The eps argument controls the factor. (The function is not online yet)

fading: This also uses Rodrigues P., et al. algorithm that in this case, uses a fading factor, also for online purposes. he eps argument controls the factor. (The function is not online yet)

Another important detail is that the standard deviation we use for all computations is the population (i.e.: divided by n), not the sample (i.e.: divided by n - 1). That is why we also provide the internally the :::std() function that computes the population, differently from stats::sd() that is the sample kind. Further more, movmean_std() shall be used when you need both results in one computation. This is faster than call mov_mean() followed by mov_std(). Finally, muinvn() is kept like that for historical reasons, as it is the function used by mpx(). It returns the sig (stable inverse centered norm) instead of std (sig is equals to 1 / (std * sqrt(window_size))).

Examples

mov <- mov_mean(motifs_discords_small, 50)

#> <simpleError in movsum_ogita_rcpp(data, window_size): object '_matrixprofiler_movsum_ogita_rcpp' not found>

mov <- mov_var(motifs_discords_small, 50)

#> <simpleError in movvar_rcpp(data, window_size): object '_matrixprofiler_movvar_rcpp' not found>

mov <- mov_sum(motifs_discords_small, 50)

#> <simpleError in movsum_ogita_rcpp(data, window_size): object '_matrixprofiler_movsum_ogita_rcpp' not found>

mov <- mov_max(motifs_discords_small, 50)

#> <simpleError in movmax_rcpp(data, window_size): object '_matrixprofiler_movmax_rcpp' not found>

mov <- mov_min(motifs_discords_small, 50)

#> <simpleError in movmin_rcpp(data, window_size): object '_matrixprofiler_movmin_rcpp' not found>

mov <- mov_std(motifs_discords_small, 50)

#> <simpleError in movstd_rcpp(data, window_size): object '_matrixprofiler_movstd_rcpp' not found>

#> Error in mov_std(motifs_discords_small, 50): object 'data_sd' not found

mov <- movmean_std(motifs_discords_small, 50)

#> Error in movmean_std_rcpp(data, window_size): object '_matrixprofiler_movmean_std_rcpp' not found

mov <- muinvn(motifs_discords_small, 50)

#> Error in muinvn_rcpp(data, window_size): object '_matrixprofiler_muinvn_rcpp' not found